# Escape HTML faster