@Kijewski 2023-03-30
The diff in the JSON escaper does not actually change anything, because the replacement is and was always 6 characters long. But I think the new version without macros is easier on the eyes.